phase4_hpe_stylegan/requirements.txt
# Deep Learning Framework
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# StyleGAN2-ADA dependencies
ninja==1.11.1
pillow==9.5.0
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
imageio==2.31.1
requests==2.31.0
tqdm==4.65.0

# HPE and similarity learning
scikit-learn==1.2.2
pandas==2.0.2
seaborn==0.12.2

# AWS integration
boto3==1.26.137
botocore==1.29.137

# Configuration and logging
pyyaml==6.0
wandb==0.15.3
tensorboard==2.13.0

# Development tools
pytest==7.3.1
black==23.3.0
flake8==6.0.0
mypy==1.3.0

# Additional utilities
click==8.1.3
rich==13.3.5
opencv-python==4.7.1.72